refactor(tlstool): make space for several splitting techniques
Part of https://github.com/ooni/probe-engine/issues/622
