+++ experiment/tlstool/internal/patternsplitter/fake_test.go
@@ -1,4 +1,4 @@
-package internal
+package patternsplitter
 
 import (
 	"context"
