+++ experiment/tlstool/internal/patternsplitter/split.go
@@ -1,4 +1,6 @@
-package internal
+// Package patternsplitter contains code to split TCP segments in
+// the middle of a specific byte pattern.
+package patternsplitter
 
 import (
 	"bytes"
@@ -9,8 +11,9 @@ import (
 	"github.com/ooni/probe-engine/netx"
 )
 
-// SplitDialer is a dialer that splits writes according to a
-// pattern and may delay the second write of delay milliseconds.
+// SplitDialer is a dialer that splits TCP segments in the middle of
+// a specific byte pattern and may optionally delay writing the second
+// half of the segment by a specific number of milliseconds.
 type SplitDialer struct {
 	netx.Dialer
 	Delay   int64
@@ -26,7 +29,13 @@ func (d SplitDialer) DialContext(ctx context.Context, network, address string) (
 	return SplitConn{Conn: conn, Delay: d.Delay, Pattern: []byte(d.Pattern)}, nil
 }
 
-// SplitConn is a conn that splits writes.
+// SplitConn is the net.Conn generated by SplitDialer.
+//
+// Caveat
+//
+// The connection will keep splitting segments when it sees the
+// specified byte pattern. This behaviour is fine as long as we're
+// just checking whether the TLS handshake works.
 type SplitConn struct {
 	net.Conn
 	BeforeSecondWrite func() // for testing
